BREAKING: Moved declaration of Actuator-related errors to the capabilities.actuator module
